+++ src/main/java/com/calories/end/dto/RecipeDTO.java
@@ -11,12 +11,12 @@ public class RecipeDTO {
     private Long id;
     private String name;
     private String description;
-    private int totalCalories;
+    private double totalCalories;
     private String translatedDescription;
     private Long userId;
     private Set<Long> ingredientIds;
 
-    public RecipeDTO(Long id, String name, String description, int totalCalories, String translatedDescription) {
+    public RecipeDTO(Long id, String name, String description, double totalCalories, String translatedDescription) {
         this.id = id;
         this.name = name;
         this.description = description;
