+++ src/main/java/com/calories/end/dto/UserDTO.java
@@ -9,6 +9,6 @@ public class UserDTO {
     private Long id;
     private String username;
     private String email;
-    private int dailyCalorieIntake;
+    private double dailyCalorieIntake;
 
 }
